Early USB: Increase the Early USB cmdline string size
The USB pid values now used are around five characters in
length. The current implementation is too restrictive w.r.t
to the USB pid length. Also, in future we may need to add
support to pass serial and product string in cmdline also.
So, to accommodate both, we doubled the size to save these
parameters.

Signed-off-by: Yadu MG <ymg@codeaurora.org>