ostree support
When we set a root= karg in the Android Boot Image, an undocumented
development mode is triggered "BootDevImage", which disables A/B
switching effectively.

In an ostree-based system, the OS should control the rootfs rather than
the bootloader. XBL/ABL should control the loading of every partition
upto and including boot_a/boot_b.